+++ src/cachelib/memcached.py
@@ -1,5 +1,7 @@
 import re
 import typing as _t
+from contextlib import nullcontext
+from functools import partial
 from time import time
 
 from cachelib.base import BaseCache
@@ -42,6 +44,16 @@ class MemcachedCache(BaseCache):
                        applications.  Keep in mind that
                        :meth:`~BaseCache.clear` will also clear keys with a
                        different prefix.
+    :param pool_size: the size of the connection pool.  This is only used if
+                      the memcached client library supports connection pooling.
+
+                      .. versionadded:: 0.15.0
+    :param pool_blocking: if the connection pool is exhausted, should the
+                          client block until a connection is available or raise
+                          an exception.  This is only used if the memcached
+                          client library supports connection pooling.
+
+                          .. versionadded:: 0.15.0
     """
 
     def __init__(
@@ -49,18 +61,26 @@ class MemcachedCache(BaseCache):
         servers: _t.Any = None,
         default_timeout: int = 300,
         key_prefix: str | None = None,
+        pool_size: int = 1,
+        pool_blocking: bool = True,
     ):
         BaseCache.__init__(self, default_timeout)
+
         if servers is None or isinstance(servers, (list, tuple)):
             if servers is None:
                 servers = ["127.0.0.1:11211"]
-            self._client = self.import_preferred_memcache_lib(servers)
-            if self._client is None:
+            client, client_context = self.import_preferred_memcache_lib(
+                servers, pool_size, pool_blocking
+            )
+            if client is None or client_context is None:
                 raise RuntimeError("no memcache module found")
+            self._client = client
+            self._client_context = client_context
         else:
             # NOTE: servers is actually an already initialized memcache
             # client.
             self._client = servers
+            self._client_context = partial(nullcontext, self._client)
 
         self.key_prefix = key_prefix
 
@@ -81,7 +101,8 @@ class MemcachedCache(BaseCache):
         # checks for so long keys can occur because it's tested from user
         # submitted data etc we fail silently for getting.
         if _test_memcached_key(key):
-            return self._client.get(key)
+            with self._client_context() as client:
+                return client.get(key)
 
     def get_dict(self, *keys: str) -> dict[str, _t.Any]:
         key_mapping = {}
@@ -90,8 +111,9 @@ class MemcachedCache(BaseCache):
             if _test_memcached_key(key):
                 key_mapping[encoded_key] = key
         _keys = list(key_mapping)
-        d: dict[str, _t.Any] = self._client.get_multi(_keys)
-        rv = d
+        with self._client_context() as client:
+            d: dict[str, _t.Any] = client.get_multi(_keys)
+            rv = d
         if self.key_prefix:
             rv = {}
             for key, value in d.items():
@@ -105,12 +127,14 @@ class MemcachedCache(BaseCache):
     def add(self, key: str, value: _t.Any, timeout: int | None = None) -> bool:
         key = self._normalize_key(key)
         timeout = self._normalize_timeout(timeout)
-        return bool(self._client.add(key, value, timeout))
+        with self._client_context() as client:
+            return bool(client.add(key, value, timeout))
 
     def set(self, key: str, value: _t.Any, timeout: int | None = None) -> bool | None:
         key = self._normalize_key(key)
         timeout = self._normalize_timeout(timeout)
-        return bool(self._client.set(key, value, timeout))
+        with self._client_context() as client:
+            return bool(client.set(key, value, timeout))
 
     def get_many(self, *keys: str) -> list[_t.Any]:
         d = self.get_dict(*keys)
@@ -125,14 +149,16 @@ class MemcachedCache(BaseCache):
             new_mapping[key] = value
 
         timeout = self._normalize_timeout(timeout)
-        failed_keys: list[_t.Any] = self._client.set_multi(new_mapping, timeout)
+        with self._client_context() as client:
+            failed_keys: list[_t.Any] = client.set_multi(new_mapping, timeout)
         k_normkey = zip(mapping.keys(), new_mapping.keys(), strict=True)
         return [k for k, nkey in k_normkey if nkey not in failed_keys]
 
     def delete(self, key: str) -> bool:
         key = self._normalize_key(key)
         if _test_memcached_key(key):
-            return bool(self._client.delete(key))
+            with self._client_context() as client:
+                return bool(client.delete(key))
         return False
 
     def delete_many(self, *keys: str) -> list[_t.Any]:
@@ -143,54 +169,69 @@ class MemcachedCache(BaseCache):
             if _test_memcached_key(normalized):
                 new_keys.append(key)
                 normalized_keys.append(normalized)
-        self._client.delete_multi(normalized_keys)
+        with self._client_context() as client:
+            client.delete_multi(normalized_keys)
         return [k for k in new_keys if not self.has(k)]
 
     def has(self, key: str) -> bool:
         key = self._normalize_key(key)
         if _test_memcached_key(key):
-            return bool(self._client.append(key, ""))
+            with self._client_context() as client:
+                return bool(client.append(key, ""))
         return False
 
     def clear(self) -> bool:
-        return bool(self._client.flush_all())
+        with self._client_context() as client:
+            return bool(client.flush_all())
 
     def inc(self, key: str, delta: int = 1) -> int | None:
         key = self._normalize_key(key)
-        value = (self._client.get(key) or 0) + delta
+        with self._client_context() as client:
+            value = (client.get(key) or 0) + delta
         return value if self.set(key, value) else None
 
     def dec(self, key: str, delta: int = 1) -> int | None:
         key = self._normalize_key(key)
-        value = (self._client.get(key) or 0) - delta
+        with self._client_context() as client:
+            value = (client.get(key) or 0) - delta
         return value if self.set(key, value) else None
 
-    def import_preferred_memcache_lib(self, servers: _t.Any) -> _t.Any:
+    def import_preferred_memcache_lib(
+        self, servers: _t.Any, pool_size: int, pool_blocking: bool = True
+    ) -> tuple[_t.Any | None, _t.Callable[[], _t.ContextManager[_t.Any]] | None]:
         """Returns an initialized memcache client.  Used by the constructor."""
         try:
             import pylibmc  # type: ignore
         except ImportError:
             pass
         else:
-            return pylibmc.Client(servers)
+            pool = pylibmc.ClientPool(pylibmc.Client(servers), pool_size)
+            reserve = partial(pool.reserve, block=pool_blocking)
+            return pool, reserve
 
         try:
             from google.appengine.api import memcache  # type: ignore
         except ImportError:
             pass
         else:
-            return memcache.Client()
+            client = memcache.Client()
+            return client, partial(nullcontext, client)
 
         try:
             import memcache  # type: ignore
         except ImportError:
             pass
         else:
-            return memcache.Client(servers)
+            client = memcache.Client(servers)
+            return client, partial(nullcontext, client)
 
         try:
             import libmc  # type: ignore
         except ImportError:
             pass
         else:
-            return libmc.Client(servers)
+            pool = libmc.ClientPool(libmc.Client(servers), pool_size)
+            reserve = partial(pool.reserve, block=pool_blocking)
+            return pool, reserve
+
+        return None, None