+++ CHANGES.rst
@@ -4,6 +4,7 @@ Version 0.15.0
 Unreleased
 
 - Make ``SimpleCache`` thread-safe using a ``threading.RLock``. :issue:`446`
+- Make ``pylibmc`` and ``libmc`` clients thread-safe by using ``ClientPool``. :pr:`287`
 - Drop support for Python 3.8, 3.9 and 3.10. :issue:`477`